third_party/WebKit/LayoutTests/platform/win/http/tests/security/contentSecurityPolicy/strict-dynamic/script-src-strict-dynamic-whitelist-expected.txt

CONSOLE WARNING: line 14: A Parser-blocking, cross site (i.e. different eTLD+1) script, http://localhost:8000/security/contentSecurityPolicy/resources/loaded.js?write, is invoked via document.write. The network request for this script MAY be blocked by the browser in this or a future page load due to poor network connectivity. If blocked in this page load, it will be confirmed in a subsequent console message.See https://www.chromestatus.com/feature/5718547946799104 for more details.
CONSOLE ERROR: line 14: Refused to load the script 'http://localhost:8000/security/contentSecurityPolicy/resources/loaded.js?write' because it violates the following Content Security Policy directive: "script-src 'nonce-abcdefg' 'strict-dynamic' http://localhost:8000". 'strict-dynamic' is present, so host-based whitelisting is disabled.

CONSOLE WARNING: line 14: A Parser-blocking, cross site (i.e. different eTLD+1) script, http://localhost:8000/security/contentSecurityPolicy/resources/loaded.js?write, is invoked via document.write. The network request for this script MAY be blocked by the browser in this or a future page load due to poor network connectivity. If blocked in this page load, it will be confirmed in a subsequent console message.See https://www.chromestatus.com/feature/5718547946799104 for more details.
CONSOLE ERROR: line 20: Refused to load the script 'http://localhost:8000/security/contentSecurityPolicy/resources/loaded.js?write-defer' because it violates the following Content Security Policy directive: "script-src 'nonce-abcdefg' 'strict-dynamic' http://localhost:8000". 'strict-dynamic' is present, so host-based whitelisting is disabled.

CONSOLE ERROR: line 26: Refused to load the script 'http://localhost:8000/security/contentSecurityPolicy/resources/loaded.js?write-async' because it violates the following Content Security Policy directive: "script-src 'nonce-abcdefg' 'strict-dynamic' http://localhost:8000". 'strict-dynamic' is present, so host-based whitelisting is disabled.

This is a testharness.js-based test.
PASS Script injected via 'document.write' is not allowed with 'strict-dynamic', even if whitelisted. 
PASS Deferred script injected via 'document.write' is not allowed with 'strict-dynamic', even if whitelisted. 
PASS Async script injected via 'document.write' is not allowed with 'strict-dynamic', even if whitelisted. 
Harness: the test ran to completion.
